Move the legacy circuit update logic to its own file outside of the ux code
