+++ npm/qsharp/src/utils.ts
@@ -1,6 +1,9 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
+import { Operation } from "./shared/circuit.js";
+import { Register } from "./shared/register.js";
+
 export type Tick = {
   value: number;
   label: string;
@@ -198,4 +201,54 @@ export function getRanges(data: SeriesOfPoints, rangeCoefficient: number) {
     max: maxY * rangeCoefficient,
   };
   return { rangeX, rangeY };
+}
+
+/**
+ * Get the minimum and maximum register indices for a given operation.
+ *
+ * @param operation The operation for which to get the register indices.
+ * @param numQubits The number of qubits in the circuit.
+ * @returns A tuple containing the minimum and maximum register indices.
+ */
+export function getMinMaxRegIdx(
+  operation: Operation,
+  numQubits: number,
+): [number, number] {
+  let targets: Register[];
+  let controls: Register[];
+  switch (operation.kind) {
+    case "measurement":
+      targets = operation.results;
+      controls = operation.qubits;
+      break;
+    case "unitary":
+      targets = operation.targets;
+      controls = operation.controls || [];
+      break;
+    case "ket":
+      targets = operation.targets;
+      controls = [];
+      break;
+  }
+
+  const ctrls: Register[] = controls || [];
+  const qRegs: Register[] = [...ctrls, ...targets].filter(
+    ({ result }) => result === undefined,
+  );
+  const qRegIdxList: number[] = qRegs.map(({ qubit }) => qubit);
+  const clsControls: Register[] = ctrls.filter(
+    ({ result }) => result !== undefined,
+  );
+  const isClassicallyControlled: boolean = clsControls.length > 0;
+  if (!isClassicallyControlled && qRegs.length === 0) return [-1, -1];
+  // If operation is classically-controlled, pad all qubit registers. Otherwise, only pad
+  // the contiguous range of registers that it covers.
+  const minRegIdx: number = isClassicallyControlled
+    ? 0
+    : Math.min(...qRegIdxList);
+  const maxRegIdx: number = isClassicallyControlled
+    ? numQubits - 1
+    : Math.max(...qRegIdxList);
+
+  return [minRegIdx, maxRegIdx];
 }
