+++ npm/qsharp/ux/circuit.tsx
@@ -5,8 +5,7 @@ import * as qviz from "./circuit-vis";
 import { useEffect, useRef, useState } from "preact/hooks";
 import { CircuitProps } from "./data.js";
 import { Spinner } from "./spinner.js";
-import { CURRENT_VERSION } from "./circuit-vis/circuit";
-import { operationListToGrid } from "./circuit-vis";
+import { toCircuitGroup } from "./circuit-vis/circuit";
 
 // For perf reasons we set a limit on how many gates/qubits
 // we attempt to render. This is still a lot higher than a human would
@@ -17,168 +16,6 @@ const MAX_QUBITS = 1000;
 
 // For now we only support one circuit at a time.
 const MAX_CIRCUITS = 1;
-
-function isCircuitGroup(circuit: any): circuit is qviz.CircuitGroup {
-  return circuit && Array.isArray(circuit.circuits);
-}
-
-function isCircuit(circuit: any): circuit is qviz.Circuit {
-  return (
-    circuit &&
-    Array.isArray(circuit.qubits) &&
-    Array.isArray(circuit.componentGrid)
-  );
-}
-
-/**
- * Get the label from a ket string.
- *
- * @param ket The ket string to extract the label from.
- * @returns The label extracted from the ket string.
- */
-function getKetLabel(ket: string): string {
-  // Check that the ket conforms to the format |{label}> or |{label}⟩
-  const ketRegex = /^\|([^\s〉⟩〉>]+)(?:[〉⟩〉>])$/;
-
-  // Match the ket string against the regex
-  const match = ket.match(ketRegex);
-
-  // If valid, return the inner label (captured group 1), otherwise return an empty string
-  return match ? match[1] : "";
-}
-
-function toOperation(op: any): qviz.Operation {
-  let targets = [];
-  if (op.targets) {
-    targets = op.targets.map((t: any) => {
-      return {
-        qubit: t.qId,
-        result: t.cId,
-      };
-    });
-  }
-  let controls = undefined;
-  if (op.controls) {
-    controls = op.controls.map((c: any) => {
-      return {
-        qubit: c.qId,
-        result: c.cId,
-      };
-    });
-  }
-
-  if (op.isMeasurement) {
-    return {
-      ...op,
-      kind: "measurement",
-      qubits: controls || [],
-      results: targets,
-    } as qviz.Operation;
-  } else {
-    const ket = getKetLabel(op.gate);
-    if (ket.length > 0) {
-      return {
-        ...op,
-        kind: "ket",
-        gate: ket,
-        targets,
-      };
-    } else {
-      const convertedOp: qviz.Operation = {
-        ...op,
-        kind: "unitary",
-        targets,
-        controls,
-      };
-      if (op.displayArgs) {
-        convertedOp.args = [op.displayArgs];
-        // Assume the parameter is always "theta" for now
-        convertedOp.params = [{ name: "theta", type: "Double" }];
-      }
-      if (op.children) {
-        convertedOp.children = [
-          {
-            components: op.children.map((child: any) => toOperation(child)),
-          },
-        ];
-      }
-      return convertedOp;
-    }
-  }
-}
-
-function toCircuitGroup(circuit: any): qviz.CircuitGroup {
-  const emptyCircuit: qviz.Circuit = {
-    qubits: [],
-    componentGrid: [],
-  };
-
-  const emptyCircuitGroup: qviz.CircuitGroup = {
-    version: CURRENT_VERSION,
-    circuits: [emptyCircuit],
-  };
-
-  if (circuit && Object.keys(circuit).length === 0) {
-    return emptyCircuitGroup;
-  }
-
-  if (circuit?.version) {
-    const version = circuit.version;
-    // If it has a "version" field, it is up-to-date
-    if (isCircuitGroup(circuit)) {
-      // If it's already a CircuitGroup, return it as is
-      return circuit;
-    } else if (isCircuit(circuit)) {
-      // If it's a Circuit, wrap it in a CircuitGroup
-      return {
-        version,
-        circuits: [circuit],
-      };
-    } else {
-      console.error(
-        "Unknown schema: circuit is neither a CircuitGroup nor a Circuit.",
-      );
-      return emptyCircuitGroup;
-    }
-  } else if (isCircuit(circuit)) {
-    // If it's a Circuit without a version, wrap it in a CircuitGroup
-    return {
-      version: CURRENT_VERSION,
-      circuits: [circuit],
-    };
-  } else if (circuit?.operations) {
-    // Legacy schema: convert to CircuitGroup
-    if (circuit.qubits === undefined || !Array.isArray(circuit.qubits)) {
-      console.error("Unknown schema: circuit is missing qubit information.");
-      return emptyCircuitGroup;
-    }
-
-    const qubits: qviz.Qubit[] = circuit.qubits.map((qubit: any) => {
-      return {
-        id: qubit.id,
-        numResults: qubit.numChildren || 0, // Rename "numChildren" to "numResults"
-      };
-    });
-
-    const componentGrid = operationListToGrid(
-      circuit.operations.map(toOperation),
-      qubits.length,
-    );
-
-    return {
-      version: CURRENT_VERSION,
-      circuits: [
-        {
-          qubits,
-          componentGrid,
-        },
-      ],
-    };
-  } else {
-    console.error("Unknown schema: circuit does not match any known format.");
-    return emptyCircuitGroup;
-  }
-}
 
 // This component is shared by the Python widget and the VS Code panel
 export function Circuit(props: {
